ref: move db operations to database package
