+++ internal/transport/task_get.go
@@ -1,13 +1,15 @@
 package transport
 
 import (
-	"database/sql"
 	"encoding/json"
 	"net/http"
+
+	"github.com/chnmk/todo-list-final-task/internal/database"
+	"github.com/chnmk/todo-list-final-task/internal/services"
 )
 
 func taskGET(w http.ResponseWriter, r *http.Request) {
-	var task Task
+	var task services.Task
 
 	id := r.FormValue("id")
 	if id == "" {
@@ -15,27 +17,9 @@ func taskGET(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	// Выполнение запроса
-	rows, err := DatabaseFile.Query("SELECT * FROM scheduler WHERE id = :id LIMIT 1", sql.Named("id", id))
+	task, err := database.GetTaskById(DatabaseFile, id)
 	if err != nil {
 		ReturnError(w, err.Error(), 500)
-		return
-	}
-
-	defer rows.Close()
-
-	// Чтение полученных данных
-	for rows.Next() {
-		err := rows.Scan(&task.Id, &task.Date, &task.Title, &task.Comment, &task.Repeat)
-		if err != nil {
-			ReturnError(w, err.Error(), 500)
-			return
-		}
-	}
-
-	// Возврат ошибки если задача не найдена
-	if task.Id == "" {
-		ReturnError(w, "задача не найдена", 500)
 		return
 	}
